separate schema logic
